+++ tests/constants/abstract.py
